refactor(glyph/dmint): Phase 2 — relocate 70 symbols into 4 submodules
All 70 top-level definitions from _legacy.py moved into their target
submodules per the plan:

  types.py     —   15 symbols   ( 330 lines)
  builders.py  —   17 symbols   ( 507 lines)
  chain.py     —   19 symbols   ( 934 lines)
  miner.py     —   19 symbols   (1413 lines)
  _legacy.py   —  pure re-exports (122 lines, down from 2977)

Dependency invariant verified one-way: types ← builders ← chain ← miner.
No cycles. The submodule __init__.py facade continues to point at
_legacy.py for now; Phase 3 will re-point it at the submodules.

143/143 non-mock dmint tests pass. The 21 mock.patch sites in
test_dmint_module.py and test_dmint_v1_mint.py still fail
(expected — Phase 4 updates them).

PLAN DEVIATION: The plan placed `_OP_STATESEPARATOR` and four
_V1_EPILOGUE_* constants in chain.py, but they are referenced by
builders.py symbols too (build_dmint_v1_code_script and friends).
Placing them in chain.py would force a builders → chain import, which
breaks the one-way dependency rule. Fix:
  - _OP_STATESEPARATOR moved to types.py (shared module-level byte
    constant, same category as _PART_B*).
  - _V1_EPILOGUE_PREFIX, _V1_EPILOGUE_ALGO_OFFSET, _V1_EPILOGUE_SUFFIX,
    _V1_EPILOGUE_LEN moved to builders.py (the V1 build helpers
    consume them); chain.py re-exports them via the existing
    chain → builders edge so any external code that imports them
    from chain.py continues to work.
Both relocations are byte-for-byte preserving. This is the same
class of fix the plan applied to DmintState (moved out of types.py
to avoid a types → chain import).
